feat: Implement season creation and default league settings

+++ src/main/kotlin/com/pokerleaguebackend/model/PlacePoint.kt
@@ -2,6 +2,7 @@ package com.pokerleaguebackend.model
 
 import com.fasterxml.jackson.annotation.JsonBackReference
 import jakarta.persistence.*
+import java.math.BigDecimal
 
 @Entity
 @Table(name = "place_points")
@@ -11,7 +12,7 @@ data class PlacePoint(
     val id: Long = 0,
 
     val place: Int,
-    val points: Int,
+    val points: BigDecimal,
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "league_settings_id")
