+++ src/main/kotlin/com/pokerleaguebackend/payload/PlacePointDto.kt
@@ -1,6 +1,8 @@
 package com.pokerleaguebackend.payload
 
+import java.math.BigDecimal
+
 data class PlacePointDto(
     val place: Int,
-    val points: Int
+    val points: BigDecimal
 )
